scripts/requirements.txt
google-cloud-firestore==0.27.0
pandas==0.21.0
xlrd==1.1.0